feat:Wallet and Order entity updated

+++ src/main/java/com/supermaket/GMarket/service/UserService.java
@@ -3,7 +3,6 @@ package com.supermaket.GMarket.service;
 import com.supermaket.GMarket.DTO.UserDTO;
 import com.supermaket.GMarket.DTO.UserOrderDTO;
 import com.supermaket.GMarket.DTO.UserProductsDTO;
-import com.supermaket.GMarket.DTO.UserWalletDTO;
 import com.supermaket.GMarket.entity.*;
 import com.supermaket.GMarket.exceptions.NotFoundException;
 import com.supermaket.GMarket.mapper.UserMapper;
@@ -113,21 +112,13 @@ public class UserService {
         }
     }
 
-    public BaseBodyResponse<UserWalletDTO> getByWallets(Long userId) {
-        Optional<User> userOptional = userRepository.findById(userId);
-        if(userOptional.isEmpty()){
-            throw new NotFoundException("Usuário com o id: " + userId + " não foi encontrado");
-        } else {
-            User user = userOptional.get();
-            List<Wallet> wallets = user.getWallets();
-            return UserMapper.toResponseWalletID(user, wallets);
-        }
-    }
+
 
     public void deleteUser(Long userId) {
         User user = userRepository.findById(userId).orElse(null);
         if (user != null) {
-            for (Wallet wallet : user.getWallets()) {
+            Wallet wallet = user.getWallet();
+            if  (wallet != null) {
                 walletRepository.delete(wallet);
             }
             Address address = user.getAddress();
